fix(e2e): extract fleet-reach loop into own file to satisfy 1000-line cap (Refs #2336)
Lift `e2eFleetReachTurnLoop`, `E2eFleetReachLoopExit`,
`E2eFleetReachLoopResult`, and `kE2eDefaultFleetReachLoopMaxTurns` from
`app/integration_test/e2e_test_shared_panels.dart` into a sibling
`e2e_test_shared_fleet_reach_loop.dart` and re-export it from the
panels barrel. Mirrors the existing
`e2e_test_shared_panel_text_assertions.dart` split pattern.

After lifting the fleet-reach turn loop in the prior commit, the
panels file grew to 1090 non-comment lines, tripping the
`repo.dart_file_non_comment_line_size` gate (≤ 1000 non-comment
lines per `SPEC/program/repo-lint.md`) and failing the
`Quality (lint, analyze, observer)` workflow on PR #2730.

The extracted file imports `e2e_test_shared.dart` (which transitively
re-exports `e2e_test_shared_panels.dart`) to reach
`e2eOpenNavalPanel`, `e2eTryNavalMoveSegment`,
`kE2eDefaultNavalMoveSegmentUiWait`, and the shared shape helpers
(`e2eFleetReachDoneFromCtSnapshotOnly`,
`e2eHarnessDetectsNonHomeFleetInNewWorld`,
`e2eNavalPanelShowsNonHomeFleetInNewWorld`, etc.). The AC1 barrel
in `e2e_helpers.dart` continues to surface
`fleetReachTurnLoop` / `E2eFleetReachLoopExit` /
`E2eFleetReachLoopResult` / `kE2eDefaultFleetReachLoopMaxTurns`
unchanged because they remain reachable through
`e2e_test_shared.dart`'s transitive export chain.

The unused `CtE2eNavalPanelSnapshot` import is removed from
`e2e_test_shared_panels.dart` after the lift (now used only by the
new file).

Verification:
- `dart run tool/ct_repo_lint.dart` -> no violations
  (`dart_file_non_comment_line_size` passes).
- `flutter analyze integration_test/ test/e2e_fleet_reach_turn_loop_test.dart`
  -> No issues found.
- `flutter test test/e2e_fleet_reach_turn_loop_test.dart
  test/e2e_helpers_barrel_test.dart` -> 41/41 pass.
- `flutter test test/e2e_test_shared_smoke_test.dart
  test/e2e_expect_panel_texts_match_snapshot_test.dart
  test/e2e_try_naval_move_segment_test.dart
  test/e2e_pick_move_destination_and_confirm_test.dart
  test/e2e_tap_move_on_first_non_home_fleet_test.dart` -> 58/58 pass.

Refs #2336
